JAMES-2082 Re-organize cassandra project hierarchy
Ids deserve their own package, Limit should be in utils
