+++ mailbox/cassandra/src/main/java/org/apache/james/mailbox/cassandra/ids/BlobId.java
@@ -16,65 +16,57 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
-package org.apache.james.mailbox.cassandra;
 
-import java.util.Objects;
-import java.util.UUID;
+package org.apache.james.mailbox.cassandra.ids;
 
-import org.apache.james.mailbox.model.MailboxId;
+import org.apache.commons.codec.digest.DigestUtils;
 
-import com.datastax.driver.core.utils.UUIDs;
+import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.MoreObjects;
+import com.google.common.base.Objects;
+import com.google.common.base.Preconditions;
+import com.google.common.base.Strings;
 
-public class CassandraId implements MailboxId {
-
-    public static class Factory implements MailboxId.Factory {
-        @Override
-        public CassandraId fromString(String serialized) {
-            return of(UUID.fromString(serialized));
-        }
-    }
-    
-    private final UUID id;
-
-    public static CassandraId timeBased() {
-        return of(UUIDs.timeBased());
+public class BlobId {
+    public static BlobId forPayload(byte[] payload) {
+        Preconditions.checkArgument(payload != null);
+        return new BlobId(DigestUtils.sha1Hex(payload));
     }
 
-    public static CassandraId of(UUID id) {
-        return new CassandraId(id);
+    public static BlobId from(String id) {
+        Preconditions.checkArgument(!Strings.isNullOrEmpty(id));
+        return new BlobId(id);
     }
 
-    private CassandraId(UUID id) {
+    private final String id;
+
+    @VisibleForTesting
+    BlobId(String id) {
         this.id = id;
     }
 
-    @Override
-    public String serialize() {
-        return id.toString();
-    }
-
-    public UUID asUuid() {
+    public String getId() {
         return id;
     }
 
     @Override
-    public final boolean equals(Object o) {
-        if (o instanceof CassandraId) {
-            CassandraId other = (CassandraId) o;
-            return Objects.equals(id, other.id);
+    public final boolean equals(Object obj) {
+        if (obj instanceof BlobId) {
+            BlobId other = (BlobId) obj;
+            return Objects.equal(id, other.id);
         }
         return false;
     }
 
     @Override
     public final int hashCode() {
-        return Objects.hash(id);
+        return Objects.hashCode(id);
     }
 
     @Override
     public String toString() {
-        return MoreObjects.toStringHelper(this)
+        return MoreObjects
+            .toStringHelper(this)
             .add("id", id)
             .toString();
     }
