+++ mailbox/cassandra/src/main/java/org/apache/james/mailbox/cassandra/ids/CassandraMailboxIdDeserializer.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.mailbox.cassandra;
+package org.apache.james.mailbox.cassandra.ids;
 
 import java.util.UUID;
